API url 수정 및 comment API 관심사 분리
