+++ src/main/java/book/book/book/api/MemberBookApi.java
@@ -9,36 +9,24 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("api/v1/memberbook/{book_id}")
+@RequestMapping("api/v1/memberbooks")
 @RequiredArgsConstructor
 public class MemberBookApi {
 
     private final MemberBookService memberBookService;
 
-    @PostMapping("/reading-status")
+    @PostMapping("/{book_id}/reading_status")
     public void saveReadingStatus(@AuthenticationPrincipal Long memberId,
                                   @PathVariable Long book_id,
                                   @RequestBody @Valid SaveReadingStatusRequest request) {
         memberBookService.saveReadingStatus(memberId, book_id, request);
     }
 
-    @DeleteMapping("/reading-status")
+    @DeleteMapping("/{book_id}/reading_status")
     public void deleteReadingStatus(@AuthenticationPrincipal Long memberId,
                                     @PathVariable Long book_id) {
         memberBookService.deleteReadingStatus(memberId, book_id);
     }
 
-    @PostMapping("/book-comment")
-    public void saveBookComment(@AuthenticationPrincipal Long memberId,
-                                @PathVariable Long book_id,
-                                @RequestBody @Valid SavedBookCommentRequest request) {
-        memberBookService.saveBookComment(memberId, book_id, request);
-    }
 
-
-    @DeleteMapping("/book-comment")
-    public void deleteBookComment(@AuthenticationPrincipal Long memberId,
-                                  @PathVariable Long book_id) {
-        memberBookService.deleteBookComment(memberId, book_id);
-    }
 }
